feat: add optimistic v3 types
