+++ crates/eth-sparse-mpt/Cargo.toml
@@ -51,7 +51,7 @@ benchmark-utils = ["dep:hash-db", "dep:triehash", "dep:flate2"]
 
 [dev-dependencies]
 criterion = { version = "0.4", features = ["html_reports"] }
-rand = { version = "0.8.5", features = ["small_rng"] }
+rand = { workspace = true, features = ["small_rng"] }
 proptest = "1.5.0"
 eth-sparse-mpt = { path = ".", features = ["benchmark-utils"] }
