+++ crates/rbuilder-primitives/Cargo.toml
@@ -22,6 +22,8 @@ alloy-rpc-types-eth.workspace = true
 revm.workspace = true
 revm-inspectors.workspace = true
 
+ethereum-consensus.workspace = true
+
 # misc
 derivative = "2.2.0"
 integer-encoding = "4.0.0"
@@ -49,3 +51,6 @@ reth-primitives.workspace = true
 reth-ethereum-primitives.workspace = true
 reth-node-core.workspace = true
 reth.workspace = true
+
+[dev-dependencies]
+rand.workspace = true
