+++ crates/rbuilder/Cargo.toml
@@ -92,10 +92,10 @@ secp256k1 = { workspace = true, features = [
 rayon = "1.8.0"
 flate2.workspace = true
 # Version required by ethereum-consensus beacon-api-client
-mev-share-sse = { git = "https://github.com/paradigmxyz/mev-share-rs", rev = "9eb2b0138ab3202b9eb3af4b19c7b3bf40b0faa8", default-features = false }
+mev-share-sse.workspace = true
+beacon-api-client.workspace = true
+ethereum-consensus.workspace = true
 jsonrpsee = { version = "0.20.3", features = ["full"] }
-beacon-api-client = { git = "https://github.com/ralexstokes/ethereum-consensus/", rev = "5031d31e318dd861cf3373702c5d92f085d926e4" }
-ethereum-consensus = { git = "https://github.com/ralexstokes/ethereum-consensus/", rev = "5031d31e318dd861cf3373702c5d92f085d926e4" }
 uuid = { version = "1.6.1", features = ["serde", "v5", "v4"] }
 prometheus.workspace = true
 warp.workspace = true
@@ -103,7 +103,7 @@ lazy_static.workspace = true
 ctor.workspace = true
 toml = { version = "0.8.8", default-features = false }
 ahash.workspace = true
-rand = "0.8.5"
+rand.workspace = true
 lru = "0.12.1"
 flume = "0.11.0"
 crossbeam-queue = "0.3.10"
