+++ crates/sysperf/Cargo.toml
@@ -10,7 +10,7 @@ repository.workspace = true
 [dependencies]
 alloy-primitives.workspace = true
 
+rand.workspace = true
 rayon = "1.8"
-rand = "0.8"
 num_cpus = "1.16"
 sysinfo = "0.33.0"
